stlye: 폴더 및 컴포넌트명 변경
