+++ src/pages/WriteCost/WriteCost.style.ts
@@ -1,19 +1,7 @@
-import styled from "styled-components";
+import styled from 'styled-components';
 
 export const WriteCostWrapper = styled.div`
-${({ theme: { mixin } }) => mixin.flexCenter({ })};
-${({ theme }) => theme.fonts.m_30_500};
-	color: ${({ theme }) => theme.colors.gray[700]};
-	width: 100%;
-	padding: 6.4rem 1.6rem;
-`;
-export const SelectBox = styled.div`
-	${({ theme: { mixin } }) => mixin.flexBox({})};
-	padding-top: 4.4rem;
-	gap: 6.4rem;
-`;
-export const ItemBox = styled.div`
-	${({ theme: { mixin } }) => mixin.flexCenter({})};
-	${({ theme }) => theme.fonts.sb_16_600};
-	gap: 1.2rem;
+  ${({ theme: { mixin }}) => mixin.flexCenter({})};
+  width: 100%;
+  padding: 1.4rem 1.6rem 3.6rem 1.6rem;
 `;
